Allow builds with debug symbols + minor fixes
(1) Allow instrument with full debug symbols.
(2) Set license path in cloverAggregateDatabases task.  Previously, the task would fail if the instrument step was up-to-date.
(3) Mark $buildDir/reports/clover as an output of the cloverGenerateReport task.  Previously, reports were not generated after a clean build when the reports directory was created by some other task.