Fix issue #83: Update plugin to support openclover and clover 4.x
Updated all code to allow use of OpenClover and Clover 4.x.
The main difference is the change in the location of the taskdef
for clover-optimized-testset Ant task.

Added user override to support running with a Clover dependency
that is not a versioned artifact. This is only necessesary if
the artifact is Clover 3.x. The default semantics use 4.x.

Updated the wrapper to use Gradle 2.14.1.
Changed all references to mavenCentral to use jcenter instead.