Fixed issue #42: maxComplexity of methodContext is not supported by plugin
Added the code contributed by phasebash in the unfinished PR#56
Completed the implementation with all supported features and
added integration test.

Added documentation for the new features and updated release notes.

Fixed the documentation and examples for the additionalSourceTests
to explain better the new facility.

Added some additional .gitignore filters for Eclipse artifacts