Fix licenseLocation is enforced even when using OpenClover
Changed all integration test projects to not provide the
licenseLocation so we get this tested.

The licenseLocation will default to $projectRootDir/clover.license
only if the file exists. If the file does not exist no errors will
be generated. However, if the version of clover used is expecting
a license file the execution will fail.