Fix [Issue 75](https://github.com/bmuschko/gradle-clover-plugin/issues/75)
Issue 75: Grails tests not executed with Clover coverage

The issue with the Grails plugins is that they use class
extension instead of the typical plugin composition model
other Gradle plugins use. As a result the code in the Clover
plugin that expects to find the GroovyPlugin applied to the
project does not work.

The solution is to add the potential plugin IDs for the current
grails-core Gradle plugins and treat them exactly like the Groovy
plugin is treated. We trigger the instrumentation of the code
just in time when the tests are about to be executed to make
sure we are using instrumented code for the tests.