Fix Issue #92: Tests fail due to differences in polish characters
The issue was a combination of problems related to the use of encoding.
The encoding selection is now applied uniformly to all compilers and
the Clover instrumentation. Additionally the correct encoding is also
applied to the default java and groovy compilers in all unit tests and
integration tests.

Proving that this works as expected can be done by forcing the default
encoding of the Gradle JVM to anything other than UTF-8. The most
reliable way to do this was to set the environment variable
`GRADLE_OPTS=-Dfile.encoding=windows-1252` before running any builds.

In order to prove the newly fixed encoding problem is really fixed
we must make the JVM use a non-default encoding. In Linux the JVM
will by default use UTF-8 which is actually the desired and selected
encoding so we cannot prove beyond doubt the problem was fixed.
With this change the JVM will use the windows-1252 encoding but the
compilers and unit tests will select UTF-8 by their own options.