Fix Issue #93: Support columns detail for Clover reports
Introduce a new closure backed configuration element in reports
following the columns nested element for clover-report. This is
implemented as a dynamic method construct so that the syntax is
easier to read.

We do not support the expression column since it is rather obsure
and would require additional syntax support to work.

Migrated several support classes to Java to prepare for the overall
Gradle 4 work which will eventually convert most of the production
code into Java from Groovy.