Fix Issue #99: compileOnly dependencies aren't properly added to classpath
Changed the mechanism used to get the compile classpath to reflect on
the currently compiled source set's own compileClasspath as documented
by the Gradle user guides. This change enables proper handling of
compile classpath in all cases and should eradicate the problem reported.
Integration test to prove this works as expected is added to java-project.

Fix Issue #89: Grails 3 project fails to find Groovyc class

Corrected the GroovyC classpath to reflect the built-in GroovyCompile task
own groovyClasspath. This preserves the ability to use a compile dependency
on groovy as well as default to the local Gradle Groovy instance.