Fix Issue #111: NPE when using Android plugin
Added defensive code tp prevent an NPE when this plugin is used
with the Android plugin. I believe this points to a problem in
the Android plugin setting the Test task classpath to a null instead
of an empty file collection as would be the correct way to do it.