Fix Issue #112: When Clover is disabled eliminate recompilation of sources during test task execution
This change eliminated the injected task actions for Test task type
when `enabled` flag is set to false. Previously we simply relied to
the OpenClover closer-setup Ant task enabled flag to disable instrumentation
but we also caused the recompilation of the classes when the enabled==false.
With this change there is no action and classes are compiled only once.