Fix two user reported issues
While these are both OpenClover limitations the changes allow for
a reasonable workaround in each case.

The sourceCompatibility function in clover-setup falls back to the
minimum supported Java version instead of the maximum causing newer
version of Java to fail in interesting ways with Java 1.3 level.

The NPE experienced when rendering report with columns might be
worked around by setting the numThreads to 1 instead of the default
2 threads.