Fix regression with cloverAggregateReports task output location
The 3.0.0 release changed the output location for cloverAggregateReports
task. This update restores the output location and retains the
correction to make the tasks cacheable by having the
cloverGenerateReport task for the root project only change the output
location to a separate directory. Previously the output of the
cloverGenerateReport task was overwritten by the cloverAggregateReports
and was useless to begin with.