Gradle 7.2 compatibility changes
Added the new Github Workflows configuration to switch this from Travis
CI to Github builds.

Adopted the new configurations for Java plugin so that Gradle scripts
are not causing failures when running with Gradle 7.x.

Fixed missing annotations for CloverMethodCotextConvention class. This
solved the primary incompatibility with Gradle 7.x.

Fixed functional tests that are using Spock Framework to use JUnit
platform setting for tests to make the tests work again.